+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Instrument configuration
 - Utilities for coordinate system conversion in `AircraftState`
+- `smol` source for packets in [MessagePack](https://msgpack.org/) or [JSON](https://www.json.org/)
 
 ### Changed